chore(ci): restore green CI — pin ruff/mypy, clear lint+type debt, target 3.11
CI had been red since before this session: the workflow installed unpinned
ruff>=0.5 / mypy>=1.10, which drifted to ruff 0.15.16 and mypy 2.1.0 whose new
rules the Phase 1-2 kernel predated (137 ruff + 59 mypy errors). Tests were
always green; this is pure tooling drift + accumulated lint/type debt.

Tooling (pyproject):
- Pin ruff==0.15.16, mypy==2.1.0 for reproducible CI.
- line-length 100 -> 120 (all violations <=118; formatting only, zero logic change).
- Ignore UP042 (enum.StrEnum is 3.11-only; (str, Enum) kept).
- Per-file E402 ignores for modules that mutate sys.path / probe optional deps
  before importing them (plane_frame, sway_check, test_report).

Ruff debt cleared (137 -> 0): import sorting (I001), unused imports (F401),
Optional[X] -> X | None and list[X] modernisation (UP), dead locals (F841),
semicolon splits (E702), zip(strict=) (B905), f-string fixes (F541).

Mypy strict debt cleared (59 -> 0): typed _combo_starting_with ->
LoadCombination (cascade fix in design.py), float() wraps for PyNite/numpy
no-any-return (axial, wind, plane_frame), dict[str, Any] for **kwargs section
loaders (library, build_saisc_sections), full type annotations on the
matplotlib diagram helpers, removed stale type: ignore comments.

Python: standardised on 3.11 (already requires-python>=3.11; CI runs 3.11).
The X | None syntax Pydantic evaluates at runtime needs 3.10+, so the prior
3.9 local workflow is retired. HANDOVER updated.

Verified: 447 passed on 3.11; ruff check . clean; mypy kernel/src tools clean.
No computed value changed (determinism, golden-file and exact-value tests pass).

+++ kernel/src/torenone_kernel/checks/axial.py
@@ -53,4 +53,4 @@ def cr_flexural(
 
     lam = slenderness * math.sqrt(fy_mpa / (math.pi**2 * _E))
     cr_n = _PHI * area_mm2 * fy_mpa * (1.0 + lam ** (2 * n)) ** (-1.0 / n)
-    return cr_n / 1_000.0   # N → kN
+    return float(cr_n / 1_000.0)   # N → kN
